# docs/api/openapi3.yaml
openapi: 3.0.1
info:
  title: sync
  description: ""
  version: 0.0.1
servers:
- url: ""
...
paths:
...
  /profiles/me:
    get:
      tags:
      - profile
      summary: Get Authenticated User Profile
      description: Get Authenticated User Profile
      operationId: GetAuthenticatedUser
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProfileResponse"
              examples:
                GetAuthenticatedUser:
                  value: "{\"userId\":\"1\",\"handle\":\"username\",\"name\":\"name\"\
                    ,\"email\":\"email@example.com\",\"bio\":\"bio\",\"profession\"\
                    :\"profession\",\"profileImageUrl\":\"https://example.com/profile-image.png\"\
                    ,\"isFollowing\":true,\"followerCount\":3,\"followingCount\":5,\"\
                    isOnboarded\":true,\"isEmailVerified\":true,\"isAuthenticatedUser\"\
                    :false,\"role\":\"USER\",\"contacts\":null}"
    patch:
      tags:
      - profile
      summary: Update Profile
      description: Update Profile
      operationId: UpdateProfile
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
            examples:
              UpdateProfile:
                value: "{\"name\":\"name\",\"handle\":\"username\",\"profileImageId\"\
                  :\"1\",\"removeProfileImage\":false,\"bio\":\"bio\",\"profession\"\
                  :\"profession\",\"contacts\":{\"custom\":\"https://example.com\"\
                  ,\"linkedin\":\"linkedinUser\",\"github\":\"githubUser\",\"instagram\"\
                  :\"instagramUser\",\"twitter\":\"twitterUser\"}}"
      responses:
        "204":
          description: "204"
  /profiles/{handle}:
    get:
      tags:
      - profile
      summary: Get Profile By Handle
      description: Get Profile By Handle
      operationId: GetProfileByHandle
      parameters:
      - name: handle
        in: path
        description: Handle
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProfileResponse"
              examples:
                GetProfileByHandle:
                  value: "{\"userId\":\"1\",\"handle\":\"username\",\"name\":\"name\"\
                    ,\"email\":\"email@example.com\",\"bio\":\"bio\",\"profession\"\
                    :\"profession\",\"profileImageUrl\":\"https://example.com/profile-image.png\"\
                    ,\"isFollowing\":true,\"followerCount\":3,\"followingCount\":5,\"\
                    isOnboarded\":true,\"isEmailVerified\":true,\"isAuthenticatedUser\"\
                    :false,\"role\":\"USER\",\"contacts\":null}"
...
  /tags/recommendations:
    get:
      tags:
      - tag
      summary: Get Tag Recommendations
      description: 팔로우할 만한 태그 목록을 추천합니다.
      operationId: GetTagRecommendations
      parameters:
      - name: type
        in: query
        description: 추천 태그 종류
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTagRecommendationsResponse"
              examples:
                GetTagRecommendations:
                  value: "{\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false},{\"id\":2,\"name\":\"spring\",\"\
                    description\":\"스프링 프레임워크 관련 태그\",\"postCount\":5,\"followerCount\"\
                    :1,\"projectHandle\":null,\"isFollowing\":true}]}"
...
components:
  schemas:
...
    GetProfileResponse:
      title: GetProfileResponse
      required:
      - bio
      - email
      - followerCount
      - followingCount
      - isAuthenticatedUser
      - isFollowing
      - name
      - profession
      - profileImageUrl
      - role
      - userId
      type: object
      properties:
        profession:
          type: string
          description: Profession
        role:
          type: string
          description: Role
        isFollowing:
          type: boolean
          description: Is Following
        bio:
          type: string
          description: Bio
        handle:
          type: string
          description: Handle
          nullable: true
        followingCount:
          type: number
          description: Following Count
        userId:
          type: string
          description: ID
        isEmailVerified:
          type: boolean
          description: "Is Email Verified (본인 프로필 조회 시에만 포함, 그 외에는 null)"
          nullable: true
        isAuthenticatedUser:
          type: boolean
          description: Is Authenticated User
        name:
          type: string
          description: Full Name
        followerCount:
          type: number
          description: Follower Count
        profileImageUrl:
          type: string
          description: Profile Image URL
        contacts:
          type: object
          properties:
            twitter:
              type: string
              description: Twitter Username
              nullable: true
            github:
              type: string
              description: GitHub Username
              nullable: true
            custom:
              type: string
              description: Custom Contact
              nullable: true
            instagram:
              type: string
              description: Instagram Username
              nullable: true
            linkedin:
              type: string
              description: LinkedIn Username
              nullable: true
          description: Contacts
        email:
          type: string
          description: E-mail
        isOnboarded:
          type: boolean
          description: "Is Onboarded (본인 프로필 조회 시에만 포함, 그 외에는 null)"
          nullable: true
    UpdateProfileRequest:
      title: UpdateProfileRequest
      type: object
      properties:
        profession:
          type: string
          description: Profession
          nullable: true
        profileImageId:
          type: string
          description: Profile Image ID
          nullable: true
        name:
          type: string
          description: Full Name
          nullable: true
        bio:
          type: string
          description: Bio
          nullable: true
        handle:
          type: string
          description: Handle
          nullable: true
        removeProfileImage:
          type: boolean
          description: Remove Profile Image
          nullable: true
        contacts:
          type: object
          properties:
            twitter:
              type: string
              description: Twitter Username
              nullable: true
            github:
              type: string
              description: GitHub Username
              nullable: true
            custom:
              type: string
              description: Custom Contact
              nullable: true
            instagram:
              type: string
              description: Instagram Username
              nullable: true
            linkedin:
              type: string
              description: LinkedIn Username
              nullable: true
          description: Contacts
...
    GetTagRecommendationsResponse:
      title: GetTagRecommendationsResponse
      required:
      - tags
      type: object
      properties:
        tags:
          type: array
          description: 태그 목록
          items:
            required:
            - description
            - followerCount
            - id
            - isFollowing
            - name
            - postCount
            type: object
            properties:
              isFollowing:
                type: boolean
                description: 요청자가 해당 태그를 팔로우하고 있는지 여부 (프로젝트 태그는 항상 false)
              name:
                type: string
                description: 태그 이름
              postCount:
                type: number
                description: 태그가 사용된 게시물 수
              description:
                type: string
                description: 태그 설명
              id:
                type: number
                description: 태그 ID
              followerCount:
                type: number
                description: 태그를 팔로우하는 사용자 수
              projectHandle:
                type: string
                description: 프로젝트 태그인 경우 해당 프로젝트의 핸들 (전역 태그인 경우 없음)
                nullable: true